# Make shouldShowDay function faster