--- a/components/marathon/schedule/List.vue
+++ b/components/marathon/schedule/List.vue
@@ -187,11 +187,15 @@ export default Vue.extend({
       if (index === 0) {
         return true;
       }
+
       // Otherwise, only show when the day transitioned
       const currentRun = new Date(this.runs![index].date);
       // We have an implicit index test for the index=0 case, so this is always safe
       const previousRun = new Date(this.runs![index - 1].date);
-      return this.$i18n.d(currentRun, 'longDate') !== this.$i18n.d(previousRun, 'longDate');
+
+      return currentRun.getDate() !== previousRun.getDate() ||
+        currentRun.getMonth() !== previousRun.getMonth() ||
+        currentRun.getFullYear() !== previousRun.getFullYear();
     },
     ...mapActions({
       getSchedule: 'api/schedule/get',